+++ Example/DBSwiftTest/Database/Passenger/Passenger.swift
@@ -26,16 +26,6 @@ class Passenger: NSManagedObject {
 		let sortDescriptor = NSSortDescriptor(key: JSON.Weight, ascending: true)
 		return (self.allBaggages.sortedArrayUsingDescriptors([sortDescriptor]) as? [Baggage] ?? [])
 	}
-
-	func entityInContext(context: NSManagedObjectContext) -> Passenger? {
-		if let
-			name = self.name,
-			age = self.age {
-				let predicate = NSPredicate(format: "%K ==[c] %@ && %K ==[c] %@", JSON.Name, name, JSON.Age, age)
-				return Passenger.MR_findFirstWithPredicate(predicate, inContext: context)
-		}
-		return nil
-	}
 }
 
 // MARK: - DKDBManager
@@ -124,7 +114,7 @@ extension Passenger {
 			age = GET_NUMBER(dictionary, JSON.Age) {
 				return NSPredicate(format: "%K ==[c] %@ && %K ==[c] %@", JSON.Name, name, JSON.Age, age)
 		}
-		return nil
+		return super.primaryPredicateWithDictionary(dictionary)
 	}
 
 	override func shouldUpdateEntityWithDictionary(dictionary: [NSObject : AnyObject]?, inContext savingContext: NSManagedObjectContext) -> Bool {